+++ src/components/MapFilters.vue
@@ -105,7 +105,7 @@
           <v-divider></v-divider>
           <v-card class="sub-card" flat>
             <v-card-text>
-              <v-subheader>Select a precipitation duration (hours):  &nbsp; {{durationValue}}</v-subheader>
+              <v-subheader>Select a precipitation duration (hours):  &nbsp; <span class="slider-values">{{durationValue}}</span></v-subheader>
               <v-slider
                 class="sliders"
                 v-model="precipDuration"
@@ -113,8 +113,8 @@
                 ticks
                 :tick-labels="durationTicks"
                 tick-size="4"
-                :max="4.5"
-                :min="0.5"
+                max="4.5"
+                min="0.5"
                 type="number"
               >
               </v-slider>
@@ -123,17 +123,17 @@
           <v-divider></v-divider>
           <v-card class="sub-card" flat>
             <v-card-text>
-              <v-subheader>Select a precipitation frequency (years):  &nbsp; {{ frequencyDisplayed }}</v-subheader>
+              <v-subheader>Select a precipitation magnitude (inches):  &nbsp; <span class="slider-values">{{ magnitudeDisplayed }}</span></v-subheader>
               <v-slider
                 class="sliders"
-                v-model="precipFrequency"
-                :step="frequencyStep"
+                v-model="precipMagnitude"
+                :step="magnitudeStep"
                 type="number"
                 ticks
-                :tick-labels="frequencyTicks"
+                :tick-labels="magnitudeTicks"
                 tick-size="4"
-                :max="9"
-                :min="1"
+                max="10"
+                min="1"
               >
               </v-slider>
             </v-card-text>
@@ -145,7 +145,7 @@
           text
           v-if="nullValue"
         >
-          No flood stage available at this duration and frequency.
+          No flood stage available at this duration and magnitude.
         </v-alert>
       </v-expansion-panel-content>
     </v-expansion-panel>
@@ -159,21 +159,46 @@
         checkbox: true,
         panel: [1],
         // durationValues: [0.5, 1, 2, 3, 4, 6, 8, 12, 24],
-        frequencyDisplayValues: [1, 2, 5, 10, 25, 50, 100, 200, 500, 1000],
+        magnitudeDisplayValues: [],
         durationSteps: [{0.5: 0.5}, {1: 1}, {1.5: 2}, {2: 3}, {2.5: 4}, {3: 6}, {3.5: 8}, {4: 12}, {4.5: 24}],
-        frequencySteps: [{1: "F1YEAR"}, {2: "F2YEAR"}, {3: "F5YEAR"}, {4: "F10YEAR"}, {5: "F25YEAR"}, {6: "F50YEAR"}, {7: "F100YEAR"}, {8: "F200YEAR"}, {9: "F500YEAR"}, {10: "F1000YEAR"}],
+        magnitudeSteps: [],
         durationStep: 0.5,
-        frequencyStep: 1,
+        magnitudeStep: 1,
         durationMin: 0.5,
-        frequencyMin: 1,
+        magnitudeMin: 1,
         durationMax: 4.5,
-        frequencyMax: 9,
+        magnitudeMax: 10,
         durationTicks: [0.5, 1, 2, 3, 4, 6, 8, 12, 24],
-        frequencyTicks: [1, 2, 5, 10, 25, 50, 100, 200, 500, 1000],
-        durationValue: 0.5,
-        frequencyValue: "F1YEAR",
-        frequencyDisplayed: 1,
-        nullValue: true,
+        magnitudeTicks: [],
+        durationValue: 24,
+        magnitudeValue: 13.1,
+        magnitudeDisplayed: 13.1,
+        currentMagnitudeIndex: 10,
+        nullValue: false,
+      }
+    },
+    mounted() {
+      this.changeMagnitude();
+    },
+    methods: {
+      changeMagnitude() {
+        let self = this;
+        let tableURL = `https://fim.wim.usgs.gov/server/rest/services/FIM_MO/Floods_v2/MapServer/3/query?where=Duration=${this.durationValue}&outFields=Magnitude&f=pjson`;
+        let httpRequest = new XMLHttpRequest();
+
+        httpRequest.onreadystatechange = function() {
+          if (this.readyState == 4 && this.status == 200) {
+              let response = JSON.parse(this.responseText);
+              self.magnitudeTicks = response.features.map(i => {return i.attributes.Magnitude});
+              self.magnitudeDisplayValues = self.magnitudeTicks;
+              let index = 0;
+              self.magnitudeSteps = response.features.map((i) => {index ++; return {[index]: i.attributes.Magnitude}});
+              self.precipMagnitude = self.currentMagnitudeIndex;
+          }
+        };
+
+        httpRequest.open("GET", tableURL, true);
+        httpRequest.send();
       }
     },
     watch: {
@@ -194,27 +219,30 @@
               if(Number(key) === value){
                 self.durationValue = durationArray[i][key];
                 self.$store.commit("getDurationValue", self.durationValue);
+                // Change magnitude options
+                self.changeMagnitude();
               }
             })
-            return this.$store.commit("getPrecipDuration", value);
+            this.$store.commit("getPrecipDuration", value);
           },
       },
-      precipFrequency: {
+      precipMagnitude: {
           get() {
-            return this.$store.state.precipFrequency;
+            return this.$store.state.precipMagnitude;
           },
           set(value) {
             let self = this;
-            let frequencyArray = JSON.parse(JSON.stringify(this.frequencySteps));
-            frequencyArray.forEach(function(obj, i) {
+            let magnitudeArray = JSON.parse(JSON.stringify(this.magnitudeSteps));
+            magnitudeArray.forEach(function(obj, i) {
               let key = Object.keys(obj);
               if(Number(key) === value){
-                self.frequencyValue = frequencyArray[i][key];
-                self.$store.commit("getFrequencyValue", self.frequencyValue);
-                self.frequencyDisplayed = self.frequencyDisplayValues[i];
+                self.magnitudeValue = magnitudeArray[i][key];
+                self.$store.commit("getMagnitudeValue", self.magnitudeValue);
+                self.magnitudeDisplayed = self.magnitudeDisplayValues[i];
+                self.currentMagnitudeIndex = value;
               }
             })
-            return this.$store.commit("getPrecipFrequency", value);
+            this.$store.commit("getPrecipMagnitude", value);
           },
       },
       // use v-model to set basemap state
@@ -223,7 +251,7 @@
           return this.$store.state.basemapState;
         },
         set(value) {
-          return this.$store.commit("getBasemapState", value);
+          this.$store.commit("getBasemapState", value);
         },
       },
       // use v-model to set moisture state
@@ -232,7 +260,7 @@
           return this.$store.state.moistureState;
         },
         set(value) {
-          return this.$store.commit("getMoistureState", value);
+          this.$store.commit("getMoistureState", value);
         },
       },
     }
@@ -249,6 +277,11 @@
 
 .v-slider--horizontal .v-slider__track-container {
   height: 5px !important;
+}
+
+.slider-values {
+  padding-left: 5px;
+  font-weight: bold;
 }
 
 .basemapThumb {